+++ src/pages/event/WriteEventInfoPage/NavigateButton/InfoNavigateButton.style.ts
